Fixed minor issues with uploading profile pictures, clean up insert sql scripts. Add logout functionality. Extended sponsorshipType enum.

+++ src/main/java/net/helloworld/aspects/UploadFileAspect.java
@@ -20,7 +20,9 @@ public class UploadFileAspect {
 
             Path source = Paths.get(filePath);
             Path target = Paths.get(newFilePath);
-            Files.copy(source, target);
+            if (!Files.exists(target)) {
+                Files.copy(source, target);
+            }
 
             Files.delete(source);
         }
